Add a demo script for the signature verification
The script is at examples/signature_verification.py. You can either run it
with the key at verif/my_public2048.pem and the signature at verif/my_signature.txt
(Lines 26-27 and 34-36 must be commented out since those are the lines that check the
key in the consensus corresponds to the one at /var/lib/tor/cached-certs)
or with the first key of the consensus (verif/public.pem) and the first signature
(verif/signature.txt).