Moved js-client to ES6 module.
These changes should not break anything when importing the bundle as
before, see 'src/lnn.js' for the compatibility.
Not all functions were converted to ES6 classes, but the individual
files (modules) should be individually importable.

Moving js-client to ES6 allows the JS code to be used for other
libraries, in particular for Lightnion's WebSocket, and pave the way
for unit-testing.